--- src/main/java/ch/heigvd/res/chill/domain/linois/PunkIPA.java
@@ -1,21 +1,0 @@
-package ch.heigvd.res.chill.domain.linois;
-
-import ch.heigvd.res.chill.domain.IProduct;
-
-import java.math.BigDecimal;
-
-public class PunkIPA implements IProduct {
-
-    public final static String NAME = "Punk IPA";
-    public final static BigDecimal PRICE = new BigDecimal(4.2);
-
-    @Override
-    public String getName() {
-        return NAME;
-    }
-
-    @Override
-    public BigDecimal getPrice() {
-        return PRICE;
-    }
-}
